+++ src/main/java/de/vinnie/mysql/profile/DataProfiles.java
@@ -3,11 +3,11 @@ package de.vinnie.mysql.profile;
 public enum DataProfiles {
 
     /*
-      _____     _     _
-     |_   _|_ _| |__ | | ___
-       | |/ _` | '_ \| |/ _ \
-       | | (_| | |_) | |  __/
-       |_|\__,_|_.__/|_|\___|
+      ____       _
+     / ___|  ___| |__   ___ _ __ ___   __ _
+     \___ \ / __| '_ \ / _ \ '_ ` _ \ / _` |
+      ___) | (__| | | |  __/ | | | | | (_| |
+     |____/ \___|_| |_|\___|_| |_| |_|\__,_|
     */
     NETWORK_PLAYER_PROFILES("network_player_profiles"),
     NETWORK_SETTINGS_PROFILES("network_settings_profiles"),
@@ -19,8 +19,8 @@ public enum DataProfiles {
        \ V / (_| | | |_| |  __/\__ \
         \_/ \__,_|_|\__,_|\___||___/
     */
-    GAMELEVEL("gamelevel"),
-    GAMERANK("gamerank"),
+    NETWORKLEVEL("networklevel"),
+    NETWORKRANK("networkrank"),
     PLAYTIME("playtime"),
     SUSHI("sushi"),
     LANGUAGE("language");